requirements.txt
pillow
numpy
matplotlib
marshmallow_dataclass
pyyaml
tensorboard
tabulate
tqdm